+++ src/Provider/Provider.js
@@ -481,7 +481,15 @@ class Provider {
       this.#sessionTimeoutCallback(req, res, next)
     })
     this.app.all(this.#invalidTokenRoute, async (req, res, next) => {
-      const errObj = JSON.parse(decodeURIComponent(req.query.err))
+      let errObj
+      /* istanbul ignore next */
+      try {
+        errObj = JSON.parse(decodeURIComponent(req.query.err))
+      } catch (err) {
+        errObj = {
+          message: 'Failed idToken validation.'
+        }
+      }
       res.locals.err = {
         status: 401,
         error: 'Unauthorized',